# target/classes/openapi/warehouse-openapi.yaml
openapi: 3.0.0
info:
  title: Warehouse API
  description: API for managing products in a warehouse
  version: 1.0.0

paths:
  /warehouse:
    get:
      summary: List all warehouses units
      responses:
        '200':
          description: A list of warehouse units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Warehouse'
    post:
      summary: Create a new warehouse unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Warehouse'
      responses:
        '201':
          description: Warehouse unit created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Warehouse'
        
        '400':
          description: Invalid request parameters

  /warehouse/search:
    get:
      summary: Search warehouses
      operationId: searchWarehouses
      parameters:
        - name: location
          in: query
          required: false
          schema:
            type: string
        - name: minCapacity
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: maxCapacity
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            default: "createdAt"
        - name: sortOrder
          in: query
          required: false
          schema:
            type: string
            default: "asc"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: A list of warehouses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Warehouse'

  /warehouse/{id}:
    get:
      summary: Get a warehouse unit by ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the warehouse unit to get
          schema:
            type: string
      responses:
        '200':
          description: Warehouse unit found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Warehouse'
        '404':
          description: Warehouse unit not found
    delete:
      summary: Archive a warehouse unit by ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the warehouse unit to archive
          schema:
            type: string
      responses:
        '204':
          description: Warehouse unit archived
        '404':
          description: Warehouse unit not found

  /warehouse/{businessUnitCode}/replacement:
    post:
      summary: Replace the current active Warehouse
      description: | 
        Replaces the current active Warehouse identified by `businessUnitCode` unit by a new Warehouse provided in the request body
        A Warehouse can be replaced by another Warehouse with the same Business Unit Code. 
        That means that the previous Warehouse will be archived and the new Warehouse will be created assuming its place.
      parameters:
        - name: businessUnitCode
          in: path
          required: true
          description: Business unit code of the warehouse unit to replace
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Warehouse'
      responses:
        '200':
          description: Warehouse unit replaced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Warehouse'
        '404':
          description: Warehouse unit not found

        '400':
          description: Invalid request parameters
components:
  schemas:
    Warehouse:
      type: object
      properties:
        id:
          type: string
          example: "456"
        businessUnitCode:
          type: string
          example: "MWH.001"
        location:
          type: string
          example: "AMSTERDAM-001"
        capacity:
          type: integer
          example: 100
        stock:
          type: integer
          example: 50